be/gestionDesNotes/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.2
djoser==2.0.5
Markdown==3.3
pytz==2022.1
sqlparse==0.4.2
